+++ src/main/java/penguin/ByeCommand.java
@@ -1,3 +1,5 @@
+package penguin;
+
 public class ByeCommand extends Command {
     public ByeCommand(String input) {
         super(input);
