+++ src/main/java/penguin/CorruptedLineException.java
@@ -1,3 +1,5 @@
+package penguin;
+
 class CorruptedLineException extends Exception {
     CorruptedLineException(String msg) {
         super(msg);
