+++ src/main/java/penguin/Deadline.java
@@ -1,3 +1,5 @@
+package penguin;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
