+++ src/main/java/penguin/Event.java
@@ -1,3 +1,5 @@
+package penguin;
+
 public class Event extends Task {
     protected String from;
     protected String to;
