+++ src/main/java/penguin/EventCommand.java
@@ -1,20 +1,17 @@
+package penguin;
+
 import java.io.IOException;
 
-public class DeleteCommand extends Command {
-    public DeleteCommand(String input) {
+public class EventCommand extends Command {
+    public EventCommand(String input) {
         super(input);
     }
 
     @Override
     public boolean execute(TaskList tasks, Ui ui, Storage storage) throws PenguinException {
-        int id = Parser.parseIndex(input, 7);
-        if (!tasks.isValidIndex(id)) {
-            ui.showBadId();
-            return false;
-        }
-
-        Task removed = tasks.removeAt(id - 1);
-        ui.showRemovedTask(removed, tasks.size());
+        Task t = Parser.parseEvent(input);
+        tasks.add(t);
+        ui.showAddedTask(t, tasks.size());
 
         try {
             storage.save(tasks.getTasks());
