+++ src/main/java/penguin/Parser.java
@@ -1,3 +1,5 @@
+package penguin;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.util.LinkedHashMap;
