+++ src/main/java/penguin/Penguin.java
@@ -1,6 +1,6 @@
+package penguin;
+
 import java.io.IOException;
-import java.time.LocalDate;
-import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
