+++ src/main/java/penguin/PenguinException.java
@@ -1,3 +1,5 @@
+package penguin;
+
 public class PenguinException extends Exception {
     public PenguinException(String message) {
         super(message);
