+++ src/main/java/penguin/Storage.java
@@ -1,3 +1,5 @@
+package penguin;
+
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
