+++ src/main/java/penguin/Task.java
@@ -1,3 +1,5 @@
+package penguin;
+
 public class Task {
     protected String description;
     protected boolean isDone;
