+++ src/main/java/penguin/Ui.java
@@ -1,3 +1,5 @@
+package penguin;
+
 import java.util.List;
 
 public class Ui {
